+++ resources/views/templates/page/_partials/favicon.blade.php
@@ -1,9 +1,27 @@
 <link rel="shortcut icon" href="/chief/favicon/favicon.ico" type="image/x-icon" />
-<link rel="apple-touch-icon" sizes="180x180" href="{{ Vite::img('favicon/apple-touch-icon.png') }}" />
-<link rel="icon" type="image/png" sizes="32x32" href="{{ Vite::img('favicon/favicon-32x32.png') }}" />
-<link rel="icon" type="image/png" sizes="16x16" href="{{ Vite::img('favicon/favicon-16x16.png') }}" />
+<link
+    rel="apple-touch-icon"
+    sizes="180x180"
+    href="{{ Vite::buildAsset('resources/assets/img/favicon/apple-touch-icon.png', 'chief/build') }}"
+/>
+<link
+    rel="icon"
+    type="image/png"
+    sizes="32x32"
+    href="{{ Vite::buildAsset('resources/assets/img/favicon/favicon-32x32.png', 'chief/build') }}"
+/>
+<link
+    rel="icon"
+    type="image/png"
+    sizes="16x16"
+    href="{{ Vite::buildAsset('resources/assets/img/favicon/favicon-16x16.png', 'chief/build') }}"
+/>
 
-<link rel="mask-icon" href="{{ Vite::img('favicon/safari-pinned-tab.svg') }}" color="#6366f1" />
-<link rel="manifest" href="{{ Vite::img('favicon/site.webmanifest') }}" />
+<link
+    rel="mask-icon"
+    href="{{ Vite::buildAsset('resources/assets/img/favicon/safari-pinned-tab.svg', 'chief/build') }}"
+    color="#6366f1"
+/>
+<link rel="manifest" href="{{ Vite::buildAsset('resources/assets/img/favicon/site.webmanifest', 'chief/build') }}" />
 
 <meta name="theme-color" content="#ffffff" />